endpoint(pipe),fix: list the same fd in a different process
+|-E option should not print a fd as an endpoint for the fd.
Consider following a bit odd example code:

  int pd[2];
  pipe(pd);
  fork();

4 file descriptors appear in this case: the parent's read and write
endpoints(fd0, fd1), and the child's read and write endpoints(fd2, fd3).

For fd0, lsof +|-E should print fd1, fd2, and fd3 as the endpoints.
However, the original code only prints fd1 and fd3 because of a bug.

fd0 (in the parent process) and fd2 (in the child process) have the
same integer value. The original code compares only the integer value
to avoid printing fd0 as the endpoint for fd0. With the logic, fd2 is
never printed.

This commit checks both the integer value and process identifiers for
the file descriptors. fd0 and fd2 have the same integer value, however
the owner processes are not the same. The new logic can distinguish
the two file descriptors.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>